requirements.txt
fastapi==0.95.0
uvicorn[standard]
gunicorn
python-dotenv==1.0.0
pandas==2.0.0
openai==0.27.0
pydantic>=1.10.8,<2.0.0
python-multipart==0.0.7
SQLAlchemy==2.0.39
aiofiles==23.2.1

# Azure dependencies
pyodbc==4.0.39
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-ai-openai==1.0.0
sqlite3-wrapper==1.0.0
alembic==1.12.0